SCRD-5556 Disable Spectre v2 (retpoline) mitigation
For nested astack deployments within some OpenStack cloud environments,
i.e. when running astack in a nested virtualisation setup, our SP4
kernels with Spectre v2 mitigation enabled fail to boot with a kernel
panic. This change disables the mitigation to avoid the problem.

This change also converts the sles12sp4 diskimage.json into a template,
and adds support automatically templating diskimage.json.j2 to create
the required diskimage.json files if not found. This will allow us to
modify specific settings depending on the deployment scenario if we
need to. Note that we include either the template or the static json
in the md5sum indicating the inputs the image build process, but not
both.

Add git ignorance of the generated sles12sp4 diskimage.json.

Eliminate some packages that we no longer need to install explicitly
in the generated image, since we are not longer using it to bring up
a build VM to build Ardana venvs.

Change-Id: I53f9d94c3aec9f687b2f62492031196a76ceae5a