SCRD-8938 Fix dev-env-install.yml libvirt setup (SCRD-8398)
Previous changes to the DevEnv setup playbooks to split out kvm setup
from libvirt inadvertently removed the triggering of the associated
libvirt playbooks.

This patch adds back in equivalent invocations of the libvirt install
and validate tasks.

This patch also simplifies the determination of the appopriate libvirt
group and service names, and specifies qemu-kvm rather than kvm as a
package dependency for libvirt.

Change-Id: I16fabce9768942c55310f1e1f2bdb29ce10153c8